+++ src/components/index.js
@@ -14,7 +14,7 @@ export {default as ConfirmEmail} from './Authentication/ConfirmEmail';
 export {default as ViewFood} from './View/ViewFood';
 export {default as ViewRestaurant} from './View/ViewRestaurant';
 
-export {default as OrderHistory} from './Order/OrderHistory';
+export {default as OrderHistory} from './Order/History/OrderHistory';
 
 export {default as Profile} from './Profile/Profile';
 export {default as ViewMenu} from './View/ViewMenu';
